+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 * Add use of `--show-hidden` option (#168)
 * Add command to generate `ignore` file (#169)
 * Use standard output (#170)
+* Fix non-displayable character, use ascii instead (#172)
 
 ## 0.8.x
 > Released Mar 08, 2023